+++ paper/src/main/java/com/telehop/paper/command/admin/TpHereAdminCommand.java
@@ -2,6 +2,7 @@ package com.telehop.paper.command.admin;
 
 import co.aikar.commands.BaseCommand;
 import co.aikar.commands.annotation.CommandAlias;
+import co.aikar.commands.annotation.CommandPermission;
 import co.aikar.commands.annotation.Default;
 import com.telehop.common.PermissionNodes;
 import com.telehop.common.model.NetworkPacket;
@@ -11,6 +12,7 @@ import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 
 @CommandAlias("tphere")
+@CommandPermission(PermissionNodes.TPHERE)
 public class TpHereAdminCommand extends BaseCommand {
     private final NetworkPaperPlugin plugin;